pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nl2sql-agents"
version = "0.1.0"
description = "A multi-agent orchestration system that converts natural language queries into safe, optimized SQL using LangGraph and OpenRouter LLMs."
readme = "README.md"
license = { text = "MIT"}
requires-python = ">=3.13"
authors = [
    { name = "Toheed Asghar" }
]
keywords = [
    "nl2sql",
    "sql",
    "natural-language",
    "multi-agent",
    "LangGraph",
    "llm",
    "text-to-sql",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "sqlparse>=0.5.0",
    "aiosqlite>=0.20.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "black>=24.0",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[project.urls]
Homepage = "https://github.com/ToheedAsghar/NL2SQL"
Repository = "https://github.com/ToheedAsghar/NL2SQL"
Issues = "https://github.com/ToheedAsghar/NL2SQL/issues"
Documentation = "https://github.com/ToheedAsghar/NL2SQL#readme"

[project.scripts]
nl2sql = "nl2sql_agents.cli:main"

# --- Tool configuration --- #

[tool.hatch.build.targets.wheel]
packages = ["nl2sql_agents"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "benchmark: Spider dataset accuracy benchmarks (require API key)",
    "slow: long-running tests (full dataset evaluations)",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false